chore: move field select item
