+++ packages/frontend/src/features/semanticViewer/components/FiltersModal/FilterInput/RelativeTimeFilter.tsx
@@ -14,7 +14,7 @@ import {
     type SelectItem,
 } from '@mantine/core';
 import { useMemo, type FC } from 'react';
-import FilterFieldSelectItem from '../FilterFieldSelectItem';
+import FilterFieldSelectItem from '../FilterFieldSelect/FilterFieldSelectItem';
 import getOperatorString from '../getOperatorString';
 
 type RelativeTimeFilterProps = {
